@@ -27,6 +27,7 @@
 //
 // *****************************************************************************
 
+#include <chrono>
 #include <mapviz/mapviz.hpp>
 
 // C++ standard libraries
@@ -34,8 +35,10 @@
 #include <cmath>
 #include <cstdlib>
 #include <fstream>
+#include <filesystem>
 #include <map>
 #include <memory>
+#include <numeric>
 #include <sstream>
 #include <string>
 #include <vector>
@@ -72,13 +75,6 @@
 
 // YAML libraries
 #include <yaml-cpp/yaml.h>
-
-// Boost libraries
-#include <boost/algorithm/string/replace.hpp>
-#include <boost/algorithm/string/split.hpp>
-#include <boost/algorithm/string/join.hpp>
-#include <boost/date_time/posix_time/posix_time.hpp>
-#include <boost/filesystem.hpp>
 
 // OpenCV libraries
 #include <opencv2/core/core.hpp>
@@ -231,6 +227,14 @@ Mapviz::Mapviz(bool is_standalone, int argc, char** argv, QWidget *parent, Qt::W
 
   ui_.bg_color->setColor(background_);
   canvas_->SetBackground(background_);
+
+  // Keyboard shortcuts for the main window
+  QShortcut * remove_display_shortcut = new QShortcut(QKeySequence(Qt::CTRL | Qt::Key_X), this);
+  connect(remove_display_shortcut, SIGNAL(activated()), this, SLOT(RemoveDisplay()));
+  QShortcut * rename_display_shortcut = new QShortcut(QKeySequence(Qt::CTRL | Qt::Key_R), this);
+  connect(rename_display_shortcut, SIGNAL(activated()), this, SLOT(RenameDisplay()));
+  QShortcut * add_display_shortcut = new QShortcut(QKeySequence(Qt::CTRL | Qt::Key_N), this);
+  connect(add_display_shortcut, SIGNAL(activated()), this, SLOT(SelectNewDisplay()));
 }
 
 Mapviz::~Mapviz()
@@ -284,9 +288,17 @@ void Mapviz::Initialize()
 
     connect(group, SIGNAL(triggered(QAction*)), this, SLOT(SetImageTransport(QAction*)));
 
+    rclcpp::QoS dynamic_tf_qos(100);
+    dynamic_tf_qos.best_effort();
+    dynamic_tf_qos.durability_volatile();
     tf_buf_ = std::make_shared<tf2_ros::Buffer>(node_->get_clock());
     tf_buf_->setUsingDedicatedThread(true);
-    tf_ = std::make_shared<tf2_ros::TransformListener>(*tf_buf_, node_, false);
+    tf_ = std::make_shared<tf2_ros::TransformListener>(
+      *tf_buf_,
+      node_,
+      false,
+      dynamic_tf_qos);
+    
     tf_manager_ = std::make_shared<swri_transform_util::TransformManager>(node_);
     try
     {
@@ -561,7 +573,7 @@ void Mapviz::Open(const std::string& filename)
 
   try
   {
-    boost::filesystem::path filepath(filename);
+    std::filesystem::path filepath(filename);
     std::string config_path = filepath.parent_path().string();
 
     ClearDisplays();
@@ -700,6 +712,25 @@ void Mapviz::Open(const std::string& filename)
           failed_plugins.push_back(type);
           RCLCPP_ERROR(node_->get_logger(), "%s", e.what());
         }
+        catch (const YAML::ParserException& e)
+        {
+          failed_plugins.push_back(type);
+          RCLCPP_ERROR(node_->get_logger(), "%s (%s) plugin failed with YAML parser error: %s",
+              type.c_str(), name.c_str(), e.what());
+          // Try to continue parsing through plugins
+        }
+        catch (const YAML::Exception& e)
+        {
+          failed_plugins.push_back(type);
+          RCLCPP_ERROR(node_->get_logger(), "%s (%s) plugin failed with YAML error: %s",
+              type.c_str(), name.c_str(), e.what());
+          // Try to continue parsing through plugins
+        }
+        catch (const rclcpp::exceptions::RCLError& e)
+        {
+          RCLCPP_ERROR(node_->get_logger(), "%s (%s) plugin failed with RCLCPP error: %s",
+              type.c_str(), name.c_str(), e.what());
+        }
       }
     }
   }
@@ -717,7 +748,15 @@ void Mapviz::Open(const std::string& filename)
   if (!failed_plugins.empty()) {
     std::stringstream message;
     message << "The following plugin(s) failed to load:" << std::endl;
-    std::string failures = boost::algorithm::join(failed_plugins, "\n");
+
+    std::string failures = std::accumulate(
+      std::next(failed_plugins.begin()),
+      failed_plugins.end(),
+      failed_plugins[0],
+      [](const std::string& a, const std::string& b) {
+        return a + "\n" + b;
+      }
+    );
     message << failures << std::endl << std::endl << "Check the ROS log for more details.";
 
     QMessageBox::warning(this, "Failed to load plugins", QString::fromStdString(message.str()));
@@ -732,7 +771,7 @@ void Mapviz::Save(const std::string& filename)
     return;
   }
 
-  boost::filesystem::path filepath(filename);
+  std::filesystem::path filepath(filename);
   std::string config_path = filepath.parent_path().string();
 
   YAML::Emitter out;
@@ -1282,12 +1321,15 @@ void Mapviz::ToggleRecord(bool on)
       AdjustWindowSize();
 
       canvas_->CaptureFrames(true);
-
-      std::string posix_time = boost::posix_time::to_iso_string(
-                                boost::posix_time::second_clock::local_time());
-      boost::replace_all(posix_time, ".", "_");
+      auto time = std::chrono::system_clock::to_time_t(std::chrono::system_clock::now());
+      std::stringstream time_stream;
+      time_stream << std::put_time(std::localtime(&time), "%Y%m%dT%H%M%S"); 
+      std::string posix_time = time_stream.str();
       std::string filename = capture_directory_ + "/mapviz_" + posix_time + ".avi";
-      boost::replace_all(filename, "~", getenv("HOME"));
+      if (filename.front() == '~')
+      {
+        filename = getenv("HOME") + filename.substr(1);
+      }
 
 
       if (!vid_writer_->initializeWriter(filename, canvas_->width(), canvas_->height())) {
@@ -1394,11 +1436,15 @@ void Mapviz::Screenshot()
 
     cv::flip(screenshot, screenshot, 0);
 
-    std::string posix_time = boost::posix_time::to_iso_string(
-                              boost::posix_time::second_clock::local_time());
-    boost::replace_all(posix_time, ".", "_");
+    auto time = std::chrono::system_clock::to_time_t(std::chrono::system_clock::now());
+    std::stringstream time_stream;
+    time_stream << std::put_time(std::localtime(&time), "%Y%m%dT%H%M%S"); 
+    std::string posix_time = time_stream.str();
     std::string filename = capture_directory_ + "/mapviz_" + posix_time + ".png";
-    boost::replace_all(filename, "~", getenv("HOME"));
+    if (filename.front() == '~')
+    {
+      filename = getenv("HOME") + filename.substr(1);
+    }
 
     RCLCPP_INFO(rclcpp::get_logger("mapviz"), "Writing screenshot to: %s", filename.c_str());
     ui_.statusbar->showMessage("Saved image to " + QString::fromStdString(filename));
@@ -1438,6 +1484,20 @@ void Mapviz::RemoveDisplay(QListWidgetItem* item)
     plugins_.erase(item);
 
     delete item;
+  }
+}
+
+void Mapviz::RenameDisplay()
+{
+  QListWidgetItem* item = ui_.configs->currentItem();
+  RenameDisplay(item);
+}
+
+void Mapviz::RenameDisplay(QListWidgetItem* item)
+{
+  if (item) {
+    ConfigItem* config_item = static_cast<ConfigItem*>(ui_.configs->itemWidget(item));
+    config_item->EditName();
   }
 }
 
