@@ -53,7 +53,7 @@ namespace mapviz
  */
 class SelectFrameDialog : public QDialog
 {
-  Q_OBJECT;
+  Q_OBJECT
 
  public:
   /**
