@@ -16,9 +16,9 @@ def generate_launch_description():
             executable="initialize_origin.py",
             name="initialize_origin",
             parameters=[
-                {"name": "local_xy_frame", "value": "map"},
-                {"name": "local_xy_origin", "value": "swri"},
-                {"name": "local_xy_origins", "value": """[
+                {"local_xy_frame": "map"},
+                {"local_xy_origin": "swri"},
+                {"local_xy_origins": """[
                     {"name": "swri",
                         "latitude": 29.45196669,
                         "longitude": -98.61370577,
@@ -29,7 +29,7 @@ def generate_launch_description():
                         "longitude": -98.629367,
                         "altitude": 200.0,
                         "heading": 0.0}
-                ]"""}
+                ]"""},
             ]
         ),
         launch_ros.actions.Node(
