@@ -33,7 +33,7 @@
 
 // ROS libraries
 #include <rclcpp/rclcpp.hpp>
-#include <tf2/transform_datatypes.h>
+#include <tf2/transform_datatypes.hpp>
 #include <gps_msgs/msg/gps_fix.hpp>
 #include <swri_transform_util/local_xy_util.h>
 
@@ -52,39 +52,41 @@ class GpsPlugin : public mapviz_plugins::PointDrawingPlugin
   Q_OBJECT
 
   public:
-  GpsPlugin();
-  ~GpsPlugin() override = default;
+    GpsPlugin();
+    ~GpsPlugin() override = default;
 
-  bool Initialize(QGLWidget* canvas) override;
-  void Shutdown() override {}
+    bool Initialize(QGLWidget* canvas) override;
+    void Shutdown() override {}
 
-  void Draw(double x, double y, double scale) override;
+    void Draw(double x, double y, double scale) override;
 
-  void LoadConfig(const YAML::Node& node, const std::string& path) override;
-  void SaveConfig(YAML::Emitter& emitter, const std::string& path) override;
+    void LoadConfig(const YAML::Node& node, const std::string& path) override;
+    void SaveConfig(YAML::Emitter& emitter, const std::string& path) override;
 
-  QWidget* GetConfigWidget(QWidget* parent) override;
+    QWidget* GetConfigWidget(QWidget* parent) override;
 
   protected:
-  void PrintError(const std::string& message) override;
-  void PrintInfo(const std::string& message) override;
-  void PrintWarning(const std::string& message) override;
+    void PrintError(const std::string& message) override;
+    void PrintInfo(const std::string& message) override;
+    void PrintWarning(const std::string& message) override;
 
   protected Q_SLOTS:
-  void SelectTopic();
-  void TopicEdited();
+    void SelectTopic();
+    void TopicEdited();
 
   private:
-  Ui::gps_config ui_;
-  QWidget* config_widget_;
+    Ui::gps_config ui_;
+    QWidget* config_widget_;
 
-  std::string topic_;
+    std::string topic_;
+    rmw_qos_profile_t qos_;
 
-  // ros::Subscriber gps_sub_;
-  rclcpp::Subscription<gps_msgs::msg::GPSFix>::SharedPtr gps_sub_;
-  bool has_message_;
+    // ros::Subscriber gps_sub_;
+    rclcpp::Subscription<gps_msgs::msg::GPSFix>::SharedPtr gps_sub_;
+    bool has_message_;
 
-  void GPSFixCallback(const gps_msgs::msg::GPSFix::SharedPtr gps);
+    void GPSFixCallback(const gps_msgs::msg::GPSFix::SharedPtr gps);
+    void connectCallback(const std::string& topic, const rmw_qos_profile_t& qos);
 };
 }   // namespace mapviz_plugins
 
