@@ -39,7 +39,7 @@
 
 // ROS libraries
 #include <rclcpp/rclcpp.hpp>
-#include <tf2/transform_datatypes.h>
+#include <tf2/transform_datatypes.hpp>
 #include <nav_msgs/msg/path.hpp>
 
 #include <mapviz/map_canvas.h>
@@ -59,38 +59,40 @@ class PathPlugin : public mapviz_plugins::PointDrawingPlugin
   Q_OBJECT
 
   public:
-  PathPlugin();
-  ~PathPlugin() override = default;
+    PathPlugin();
+    ~PathPlugin() override = default;
 
-  bool Initialize(QGLWidget* canvas) override;
-  void Shutdown() override {}
+    bool Initialize(QGLWidget* canvas) override;
+    void Shutdown() override {}
 
-  void Draw(double x, double y, double scale) override;
+    void Draw(double x, double y, double scale) override;
 
-  void LoadConfig(const YAML::Node& node, const std::string& path) override;
-  void SaveConfig(YAML::Emitter& emitter, const std::string& path) override;
+    void LoadConfig(const YAML::Node& node, const std::string& path) override;
+    void SaveConfig(YAML::Emitter& emitter, const std::string& path) override;
 
-  QWidget* GetConfigWidget(QWidget* parent) override;
+    QWidget* GetConfigWidget(QWidget* parent) override;
 
   protected:
-  void PrintError(const std::string& message) override;
-  void PrintInfo(const std::string& message) override;
-  void PrintWarning(const std::string& message) override;
+    void PrintError(const std::string& message) override;
+    void PrintInfo(const std::string& message) override;
+    void PrintWarning(const std::string& message) override;
 
   protected Q_SLOTS:
-  void SelectTopic();
-  void TopicEdited();
-
+    void SelectTopic();
+    void TopicEdited();
+  
   private:
-  Ui::path_config ui_;
-  QWidget* config_widget_;
+    Ui::path_config ui_;
+    QWidget* config_widget_;
 
-  std::string topic_;
+    std::string topic_;
+    rmw_qos_profile_t qos_;
 
-  rclcpp::Subscription<nav_msgs::msg::Path>::SharedPtr path_sub_;
-  bool has_message_;
+    rclcpp::Subscription<nav_msgs::msg::Path>::SharedPtr path_sub_;
+    bool has_message_;
 
-  void pathCallback(const nav_msgs::msg::Path::SharedPtr path);
+    void connectCallback(const std::string& topic, const rmw_qos_profile_t& qos);
+    void pathCallback(const nav_msgs::msg::Path::SharedPtr path);
 };
 }   // namespace mapviz_plugins
 
