@@ -48,7 +48,6 @@
 
 // QT autogenerated files
 #include "ui_point_click_publisher_config.h"
-#include "ui_topic_select.h"
 
 /**
  * This is a pretty straightforward plugin.  It watches for user clicks on the
@@ -59,7 +58,7 @@ namespace mapviz_plugins
 {
 class PointClickPublisherPlugin : public mapviz::MapvizPlugin
 {
-  Q_OBJECT;
+  Q_OBJECT
 public:
   PointClickPublisherPlugin();
   ~PointClickPublisherPlugin() override;
