@@ -39,7 +39,7 @@
 
 // ROS libraries
 #include <rclcpp/rclcpp.hpp>
-#include <tf2/transform_datatypes.h>
+#include <tf2/transform_datatypes.hpp>
 
 #include <mapviz/map_canvas.h>
 
@@ -48,7 +48,6 @@
 
 // QT autogenerated files
 #include "ui_robot_image_config.h"
-#include "ui_topic_select.h"
 
 namespace mapviz_plugins
 {
