@@ -39,7 +39,7 @@
 
 // ROS libraries
 #include <rclcpp/rclcpp.hpp>
-#include <tf2/transform_datatypes.h>
+#include <tf2/transform_datatypes.hpp>
 #include <marti_nav_msgs/msg/route_position.hpp>
 #include <mapviz/map_canvas.h>
 #include <swri_route_util/route.h>
@@ -59,61 +59,65 @@ class RoutePlugin : public mapviz::MapvizPlugin
   Q_OBJECT
 
   public:
-  enum DrawStyle
-  {
-    LINES = 0,
-    POINTS = 1
-  };
+    enum DrawStyle
+    {
+      LINES = 0,
+      POINTS = 1
+    };
 
-  RoutePlugin();
-  ~RoutePlugin() override = default;
+    RoutePlugin();
+    ~RoutePlugin() override = default;
 
-  bool Initialize(QGLWidget* canvas) override;
-  void Shutdown() override {}
+    bool Initialize(QGLWidget* canvas) override;
+    void Shutdown() override {}
 
-  void Draw(double x, double y, double scale) override;
+    void Draw(double x, double y, double scale) override;
 
-  void Transform() override {}
+    void Transform() override {}
 
-  void LoadConfig(const YAML::Node& node, const std::string& path) override;
-  void SaveConfig(YAML::Emitter& emitter, const std::string& path) override;
-  void DrawStopWaypoint(double x, double y);
-  void DrawRoute(const swri_route_util::Route &route);
-  void DrawRoutePoint(const swri_route_util::RoutePoint &point);
+    void LoadConfig(const YAML::Node& node, const std::string& path) override;
+    void SaveConfig(YAML::Emitter& emitter, const std::string& path) override;
+    void DrawStopWaypoint(double x, double y);
+    void DrawRoute(const swri_route_util::Route &route);
+    void DrawRoutePoint(const swri_route_util::RoutePoint &point);
 
-  QWidget* GetConfigWidget(QWidget* parent) override;
+    QWidget* GetConfigWidget(QWidget* parent) override;
 
   protected:
-  void PrintError(const std::string& message) override;
-  void PrintInfo(const std::string& message) override;
-  void PrintWarning(const std::string& message) override;
+    void PrintError(const std::string& message) override;
+    void PrintInfo(const std::string& message) override;
+    void PrintWarning(const std::string& message) override;
 
   protected Q_SLOTS:
-  void SelectTopic();
-  void SelectPositionTopic();
-  void TopicEdited();
-  void PositionTopicEdited();
-  void SetDrawStyle(QString style);
-  void DrawIcon() override;
+    void SelectTopic();
+    void SelectPositionTopic();
+    void TopicEdited();
+    void PositionTopicEdited();
+    void SetDrawStyle(QString style);
+    void DrawIcon() override;
 
   private:
-  Ui::route_config ui_;
-  QWidget* config_widget_;
+    Ui::route_config ui_;
+    QWidget* config_widget_;
 
-  DrawStyle draw_style_;
+    DrawStyle draw_style_;
 
-  std::string topic_;
-  std::string position_topic_;
+    std::string topic_;
+    rmw_qos_profile_t qos_;
+    std::string position_topic_;
+    rmw_qos_profile_t position_qos_;
 
-  rclcpp::Subscription<marti_nav_msgs::msg::Route>::SharedPtr route_sub_;
-  rclcpp::Subscription<marti_nav_msgs::msg::RoutePosition>::SharedPtr position_sub_;
+    rclcpp::Subscription<marti_nav_msgs::msg::Route>::SharedPtr route_sub_;
+    rclcpp::Subscription<marti_nav_msgs::msg::RoutePosition>::SharedPtr position_sub_;
 
-  swri_route_util::Route src_route_;
-  // marti_nav_msgs::RoutePositionConstPtr src_route_position_;
-  marti_nav_msgs::msg::RoutePosition::SharedPtr src_route_position_;
+    swri_route_util::Route src_route_;
+    // marti_nav_msgs::RoutePositionConstPtr src_route_position_;
+    marti_nav_msgs::msg::RoutePosition::SharedPtr src_route_position_;
 
-  void RouteCallback(const marti_nav_msgs::msg::Route::SharedPtr msg);
-  void PositionCallback(const marti_nav_msgs::msg::RoutePosition::SharedPtr msg);
+    void connectRouteCallback(const std::string& topic, const rmw_qos_profile_t& qos);
+    void connectPositionCallback(const std::string& topic, const rmw_qos_profile_t& qos);
+    void RouteCallback(const marti_nav_msgs::msg::Route::SharedPtr msg);
+    void PositionCallback(const marti_nav_msgs::msg::RoutePosition::SharedPtr msg);
 };
 }   // namespace mapviz_plugins
 
