@@ -18,6 +18,7 @@
 // *****************************************************************************
 
 #include <mapviz_plugins/gps_plugin.h>
+#include <mapviz_plugins/topic_select.h>
 
 // QT libraries
 #include <QDialog>
@@ -31,7 +32,6 @@
 
 #include <swri_image_util/geometry_util.h>
 #include <swri_transform_util/transform_util.h>
-#include <mapviz/select_topic_dialog.h>
 
 // Declare plugin
 #include <pluginlib/class_list_macros.hpp>
@@ -46,11 +46,13 @@ PLUGINLIB_EXPORT_CLASS(mapviz_plugins::GpsPlugin, mapviz::MapvizPlugin)
 
 namespace mapviz_plugins
 {
-  GpsPlugin::GpsPlugin()
-  : PointDrawingPlugin()
-  , ui_()
-  , config_widget_(new QWidget())
-  , has_message_(false)
+  GpsPlugin::GpsPlugin() :
+    PointDrawingPlugin(),
+    ui_(),
+    config_widget_(new QWidget()),
+    has_message_(false),
+    topic_(""),
+    qos_(rmw_qos_profile_default)
   {
     ui_.setupUi(config_widget_);
 
@@ -90,19 +92,26 @@ namespace mapviz_plugins
 
   void GpsPlugin::SelectTopic()
   {
-    std::string topic = mapviz::SelectTopicDialog::selectTopic(node_, "gps_msgs/msg/GPSFix");
-
+    auto [topic, qos] = SelectTopicDialog::selectTopic(
+      node_,
+      "gps_msgs/msg/GPSFix",
+      qos_);
     if (!topic.empty())
     {
-      ui_.topic->setText(QString::fromStdString(topic));
-      TopicEdited();
+      connectCallback(topic, qos);
     }
   }
 
   void GpsPlugin::TopicEdited()
   {
     std::string topic = ui_.topic->text().trimmed().toStdString();
-    if (topic != topic_)
+    connectCallback(topic, qos_);
+  }
+
+  void GpsPlugin::connectCallback(const std::string& topic, const rmw_qos_profile_t& qos)
+  {
+    ui_.topic->setText(QString::fromStdString(topic));
+    if ((topic != topic_) || !qosEqual(qos, qos_) )
     {
       initialized_ = false;
       ClearPoints();
@@ -112,9 +121,12 @@ namespace mapviz_plugins
       gps_sub_.reset();
 
       topic_ = topic;
+      qos_ = qos;
       if (!topic.empty())
       {
-        gps_sub_ = node_->create_subscription<gps_msgs::msg::GPSFix>(topic_, rclcpp::QoS(1),
+        gps_sub_ = node_->create_subscription<gps_msgs::msg::GPSFix>(
+          topic_,
+          rclcpp::QoS(rclcpp::QoSInitialization::from_rmw(qos), qos),
           std::bind(&GpsPlugin::GPSFixCallback, this, std::placeholders::_1));
 
         RCLCPP_INFO(node_->get_logger(), "Subscribing to %s", topic_.c_str());
@@ -194,6 +206,7 @@ namespace mapviz_plugins
 
   void GpsPlugin::LoadConfig(const YAML::Node& node, const std::string& path)
   {
+    LoadQosConfig(node, qos_);
     if (node["topic"])
     {
       std::string topic = node["topic"].as<std::string>();
@@ -288,5 +301,7 @@ namespace mapviz_plugins
       << ui_.static_arrow_sizes->isChecked();
 
     emitter << YAML::Key << "arrow_size" << YAML::Value << ui_.arrow_size->value();
+
+    SaveQosConfig(emitter, qos_);
   }
 }   // namespace mapviz_plugins
