@@ -28,9 +28,7 @@
 // *****************************************************************************
 
 #include <mapviz_plugins/laserscan_plugin.h>
-
-// Boost libraries
-#include <boost/algorithm/string.hpp>
+#include <mapviz_plugins/topic_select.h>
 
 // QT libraries
 #include <QDialog>
@@ -40,38 +38,35 @@
 #include <rclcpp/rclcpp.hpp>
 #include <swri_transform_util/transform.h>
 
-#include <mapviz/select_topic_dialog.h>
-
 // Declare plugin
 #include <pluginlib/class_list_macros.hpp>
 
 // C++ standard libraries
+#include <algorithm>
 #include <cmath>
 #include <cstdio>
 #include <deque>
 #include <string>
 
-// QT Autogenerated
-#include "ui_topic_select.h"
-
 PLUGINLIB_EXPORT_CLASS(mapviz_plugins::LaserScanPlugin, mapviz::MapvizPlugin)
 
 namespace mapviz_plugins
 {
-  LaserScanPlugin::LaserScanPlugin()
-  : MapvizPlugin()
-  , ui_()
-  , config_widget_(new QWidget())
-  , topic_("")
-  , alpha_(1.0)
-  , min_value_(0.0)
-  , max_value_(100.0)
-  , point_size_(3)
-  , buffer_size_(0)
-  , has_message_(false)
-  , prev_ranges_size_(0)
-  , prev_angle_min_(0.0)
-  , prev_increment_(0.0)
+  LaserScanPlugin::LaserScanPlugin() :
+    MapvizPlugin(),
+    ui_(),
+    config_widget_(new QWidget()),
+    topic_(""),
+    alpha_(1.0),
+    min_value_(0.0),
+    max_value_(100.0),
+    point_size_(3),
+    buffer_size_(0),
+    has_message_(false),
+    prev_ranges_size_(0),
+    prev_angle_min_(0.0),
+    prev_increment_(0.0),
+    qos_(rmw_qos_profile_default)
   {
     ui_.setupUi(config_widget_);
 
@@ -257,21 +252,26 @@ namespace mapviz_plugins
 
   void LaserScanPlugin::SelectTopic()
   {
-    std::string topic = mapviz::SelectTopicDialog::selectTopic(
+    auto [topic, qos] = SelectTopicDialog::selectTopic(
       node_,
-      "sensor_msgs/msg/LaserScan");
-
+      "sensor_msgs/msg/LaserScan",
+      qos_);
     if (!topic.empty())
     {
-      ui_.topic->setText(QString::fromStdString(topic));
-      TopicEdited();
+      connectCallback(topic, qos);
     }
   }
 
   void LaserScanPlugin::TopicEdited()
   {
     std::string topic = ui_.topic->text().trimmed().toStdString();
-    if (topic != topic_)
+    connectCallback(topic, qos_);
+  }
+
+  void LaserScanPlugin::connectCallback(const std::string& topic, const rmw_qos_profile_t& qos)
+  {
+    ui_.topic->setText(QString::fromStdString(topic));
+    if ((topic != topic_) || !qosEqual(qos, qos_))
     {
       initialized_ = false;
       scans_.clear();
@@ -281,11 +281,12 @@ namespace mapviz_plugins
       laserscan_sub_.reset();
 
       topic_ = topic;
+      qos_ = qos;
       if (!topic.empty())
       {
         laserscan_sub_ = node_->create_subscription<sensor_msgs::msg::LaserScan>(
           topic_,
-          rclcpp::QoS(100),
+          rclcpp::QoS(rclcpp::QoSInitialization::from_rmw(qos), qos),
           std::bind(&LaserScanPlugin::laserScanCallback, this, std::placeholders::_1)
         );
 
@@ -536,10 +537,11 @@ namespace mapviz_plugins
   void LaserScanPlugin::LoadConfig(const YAML::Node& node,
       const std::string& path)
   {
+    LoadQosConfig(node, qos_);
     if (node["topic"])
     {
-      std::string topic = node["topic"].as<std::string>();
-      ui_.topic->setText(boost::trim_copy(topic).c_str());
+      std::string trimmed = TrimString(node["topic"].as<std::string>());
+      ui_.topic->setText(trimmed.c_str());
       TopicEdited();
     }
 
@@ -662,8 +664,10 @@ namespace mapviz_plugins
   void LaserScanPlugin::SaveConfig(YAML::Emitter& emitter,
       const std::string& path)
   {
+    std::string trimmed = TrimString(ui_.topic->text().toStdString());
+
     emitter << YAML::Key << "topic" <<
-               YAML::Value << boost::trim_copy(ui_.topic->text().toStdString());
+               YAML::Value << trimmed;
     emitter << YAML::Key << "size" <<
                YAML::Value << ui_.pointSize->value();
     emitter << YAML::Key << "buffer_size" <<
@@ -682,6 +686,7 @@ namespace mapviz_plugins
                YAML::Value << ui_.maxValue->text().toDouble();
     emitter << YAML::Key << "use_rainbow" <<
                YAML::Value << ui_.use_rainbow->isChecked();
+    SaveQosConfig(emitter, qos_);
   }
 }   // namespace mapviz_plugins
 
