@@ -29,7 +29,7 @@
 
 #include <mapviz_plugins/point_click_publisher_plugin.h>
 #include <swri_transform_util/frames.h>
-#include <tf2/transform_datatypes.h>
+#include <tf2/transform_datatypes.hpp>
 
 // Declare plugin
 #include <pluginlib/class_list_macros.hpp>
