@@ -29,9 +29,7 @@
 
 #include <GL/glew.h>
 #include <mapviz_plugins/pointcloud2_plugin.h>
-
-// Boost libraries
-#include <boost/algorithm/string.hpp>
+#include <mapviz_plugins/topic_select.h>
 
 // QT libraries
 #include <QDialog>
@@ -40,8 +38,6 @@
 // ROS libraries
 #include <rclcpp/rclcpp.hpp>
 #include <swri_transform_util/transform.h>
-
-#include <mapviz/select_topic_dialog.h>
 
 // Declare plugin
 #include <pluginlib/class_list_macros.hpp>
@@ -53,28 +49,26 @@
 #include <utility>
 #include <vector>
 
-// QT Autogenerated
-#include "ui_topic_select.h"
-
 PLUGINLIB_EXPORT_CLASS(mapviz_plugins::PointCloud2Plugin, mapviz::MapvizPlugin)
 
 namespace mapviz_plugins
 {
-  PointCloud2Plugin::PointCloud2Plugin()
-  : MapvizPlugin()
-  , ui_()
-  , config_widget_(new QWidget())
-  , topic_("")
-  , alpha_(1.0)
-  , max_value_(100.0)
-  , min_value_(0.0)
-  , point_size_(3)
-  , buffer_size_(1)
-  , new_topic_(true)
-  , has_message_(false)
-  , num_of_feats_(0)
-  , need_new_list_(true)
-  , need_minmax_(false)
+  PointCloud2Plugin::PointCloud2Plugin() :
+    MapvizPlugin(),
+    ui_(),
+    config_widget_(new QWidget()),
+    topic_(""),
+    alpha_(1.0),
+    max_value_(100.0),
+    min_value_(0.0),
+    point_size_(3),
+    buffer_size_(1),
+    new_topic_(true),
+    has_message_(false),
+    num_of_feats_(0),
+    need_new_list_(true),
+    need_minmax_(false),
+    qos_(rmw_qos_profile_default)
   {
     ui_.setupUi(config_widget_);
 
@@ -233,7 +227,7 @@ namespace mapviz_plugins
     {
       pc2_sub_ = node_->create_subscription<sensor_msgs::msg::PointCloud2>(
         topic_,
-        rclcpp::QoS(10),
+        rclcpp::QoS(rclcpp::QoSInitialization::from_rmw(qos_)),
         std::bind(&PointCloud2Plugin::PointCloud2Callback, this, std::placeholders::_1)
       );
       new_topic_ = true;
@@ -341,14 +335,13 @@ namespace mapviz_plugins
 
   void PointCloud2Plugin::SelectTopic()
   {
-    std::string topic = mapviz::SelectTopicDialog::selectTopic(
+    auto [topic, qos] = SelectTopicDialog::selectTopic(
       node_,
-      "sensor_msgs/msg/PointCloud2");
-
+      "sensor_msgs/msg/PointCloud2",
+      qos_);
     if (!topic.empty())
     {
-      ui_.topic->setText(QString::fromStdString(topic));
-      TopicEdited();
+      connectCallback(topic, qos);
     }
   }
 
@@ -356,7 +349,13 @@ namespace mapviz_plugins
   void PointCloud2Plugin::TopicEdited()
   {
     std::string topic = ui_.topic->text().trimmed().toStdString();
-    if (topic != topic_)
+    connectCallback(topic, qos_);
+  }
+
+  void PointCloud2Plugin::connectCallback(const std::string& topic, const rmw_qos_profile_t& qos)
+  {
+    ui_.topic->setText(QString::fromStdString(topic));
+    if ((topic != topic_) || !qosEqual(qos, qos_))
     {
       initialized_ = false;
       {
@@ -367,8 +366,10 @@ namespace mapviz_plugins
       PrintWarning("No messages received.");
 
       topic_ = topic;
+      qos_ = qos;
       SetSubscription(this->Visible());
     }
+
   }
 
   void PointCloud2Plugin::MinValueChanged(double value)
@@ -733,10 +734,12 @@ namespace mapviz_plugins
   void PointCloud2Plugin::LoadConfig(const YAML::Node& node,
                                      const std::string& path)
   {
+    LoadQosConfig(node, qos_);
+
     if (node["topic"])
     {
-      std::string topic = node["topic"].as<std::string>();
-      ui_.topic->setText(boost::trim_copy(topic).c_str());
+      std::string topic = TrimString(node["topic"].as<std::string>());
+      ui_.topic->setText(topic.c_str());
       TopicEdited();
     }
 
@@ -856,12 +859,13 @@ namespace mapviz_plugins
   {
     alpha_ = std::max(0.0f, std::min(static_cast<float>(value), 1.0f));
   }
-
+  
   void PointCloud2Plugin::SaveConfig(YAML::Emitter& emitter,
                                      const std::string& path)
   {
+    std::string topic = TrimString(ui_.topic->text().toStdString());
     emitter << YAML::Key << "topic" <<
-      YAML::Value << boost::trim_copy(ui_.topic->text().toStdString());
+      YAML::Value << topic;
     emitter << YAML::Key << "size" <<
       YAML::Value << ui_.pointSize->value();
     emitter << YAML::Key << "buffer_size" <<
@@ -884,5 +888,7 @@ namespace mapviz_plugins
       YAML::Value << ui_.use_automaxmin->isChecked();
     emitter << YAML::Key << "unpack_rgb" <<
       YAML::Value << ui_.unpack_rgb->isChecked();
+
+    SaveQosConfig(emitter, qos_);
   }
 }   // namespace mapviz_plugins
