@@ -29,8 +29,7 @@
 // *****************************************************************************
 
 #include <mapviz_plugins/string_plugin.h>
-
-#include <mapviz/select_topic_dialog.h>
+#include <mapviz_plugins/topic_select.h>
 
 #include <QFontDialog>
 #include <QFontDatabase>
@@ -51,18 +50,19 @@ namespace mapviz_plugins
   const char* StringPlugin::OFFSET_Y_KEY = "offset_y";
   static constexpr int DEFAULT_FONT_SIZE = 9;
 
-  StringPlugin::StringPlugin()
-  : MapvizPlugin()
-  , ui_()
-  , config_widget_(new QWidget())
-  , anchor_(TOP_LEFT)
-  , units_(PIXELS)
-  , offset_x_(0)
-  , offset_y_(0)
-  , has_message_(false)
-  , has_painted_(false)
-  , color_(Qt::black)
-  , font_()
+  StringPlugin::StringPlugin() :
+    MapvizPlugin(),
+    ui_(),
+    config_widget_(new QWidget()),
+    anchor_(TOP_LEFT),
+    units_(PIXELS),
+    offset_x_(0),
+    offset_y_(0),
+    has_message_(false),
+    has_painted_(false),
+    color_(Qt::black),
+    font_(),
+    qos_(rmw_qos_profile_default)
   {
     ui_.setupUi(config_widget_);
     // Set background white
@@ -202,6 +202,7 @@ namespace mapviz_plugins
 
   void StringPlugin::LoadConfig(const YAML::Node& node, const std::string& path)
   {
+    LoadQosConfig(node, qos_);
     if (node[TOPIC_KEY])
     {
       ui_.topic->setText(QString(node[TOPIC_KEY].as<std::string>().c_str()));
@@ -275,6 +276,7 @@ namespace mapviz_plugins
     emitter << YAML::Key << UNITS_KEY << YAML::Value << UnitsToString(units_);
     emitter << YAML::Key << OFFSET_X_KEY << YAML::Value << offset_x_;
     emitter << YAML::Key << OFFSET_Y_KEY << YAML::Value << offset_y_;
+    SaveQosConfig(emitter, qos_);
   }
 
   QWidget* StringPlugin::GetConfigWidget(QWidget* parent)
@@ -322,21 +324,27 @@ namespace mapviz_plugins
 
   void StringPlugin::SelectTopic()
   {
-    std::string topic = mapviz::SelectTopicDialog::selectTopic(
-      node_, "std_msgs/msg/String"
-    );
+    auto [topic, qos] = SelectTopicDialog::selectTopic(
+      node_,
+      "std_msgs/msg/String",
+      qos_);
 
     if (!topic.empty())
     {
-      ui_.topic->setText(QString::fromStdString(topic));
-      TopicEdited();
+      connectCallback(topic, qos);
     }
   }
 
   void StringPlugin::TopicEdited()
   {
     std::string topic = ui_.topic->text().trimmed().toStdString();
-    if (topic != topic_)
+    connectCallback(topic, qos_);
+  }
+
+  void StringPlugin::connectCallback(const std::string& topic, const rmw_qos_profile_t& qos)
+  {
+    ui_.topic->setText(QString::fromStdString(topic));
+    if ((topic != topic_) || !qosEqual(qos, qos_))
     {
       initialized_ = false;
       has_message_ = false;
@@ -345,12 +353,13 @@ namespace mapviz_plugins
       string_sub_.reset();
 
       topic_ = topic;
+      qos_ = qos;
       if (!topic.empty())
       {
         try
         {
           string_sub_ = node_->create_subscription<std_msgs::msg::String>(topic_,
-              rclcpp::QoS(1),
+            rclcpp::QoS(rclcpp::QoSInitialization::from_rmw(qos), qos),
               [this](const std_msgs::msg::String::ConstSharedPtr str) {
             SetText(QString(str->data.c_str()));
           });
@@ -365,7 +374,7 @@ namespace mapviz_plugins
         try
         {
           string_stamped_sub_ = node_->create_subscription<marti_common_msgs::msg::StringStamped>(topic_,
-              rclcpp::QoS(1),
+            rclcpp::QoS(rclcpp::QoSInitialization::from_rmw(qos), qos),
               [this](const marti_common_msgs::msg::StringStamped::ConstSharedPtr str) {
             SetText(QString(str->value.c_str()));
           });
