@@ -28,9 +28,7 @@
 // *****************************************************************************
 
 #include <mapviz_plugins/textured_marker_plugin.h>
-
-// Boost libraries
-#include <boost/algorithm/string.hpp>
+#include <mapviz_plugins/topic_select.h>
 
 // QT libraries
 #include <QDialog>
@@ -39,8 +37,6 @@
 // ROS libraries
 #include <rclcpp/rclcpp.hpp>
 #include <sensor_msgs/image_encodings.hpp>
-
-#include <mapviz/select_topic_dialog.h>
 
 // Declare plugin
 #include <pluginlib/class_list_macros.hpp>
@@ -60,12 +56,14 @@ PLUGINLIB_EXPORT_CLASS(mapviz_plugins::TexturedMarkerPlugin, mapviz::MapvizPlugi
 
 namespace mapviz_plugins
 {
-TexturedMarkerPlugin::TexturedMarkerPlugin()
-: MapvizPlugin()
-, ui_()
-, alphaVal_(1.0f)
-, config_widget_(new QWidget())
-, has_message_(false)
+TexturedMarkerPlugin::TexturedMarkerPlugin() :
+  MapvizPlugin(),
+  ui_(),
+  alphaVal_(1.0f),
+  config_widget_(new QWidget()),
+  has_message_(false),
+  topic_(""),
+  qos_(rmw_qos_profile_default)
 {
   ui_.setupUi(config_widget_);
 
@@ -124,23 +122,28 @@ void TexturedMarkerPlugin::SetAlphaLevel(int alpha)
 
 void TexturedMarkerPlugin::SelectTopic()
 {
-  std::string topic = mapviz::SelectTopicDialog::selectTopic(
+  auto [topic, qos] = SelectTopicDialog::selectTopic(
     node_,
     "marti_visualization_msgs/msg/TexturedMarker",
-    "marti_visualization_msgs/msg/TexturedMarkerArray"
-  );
+    "marti_visualization_msgs/msg/TexturedMarkerArray",
+    qos_);
 
   if (!topic.empty()) {
-    ui_.topic->setText(QString::fromStdString(topic));
-
-    TopicEdited();
+    connectCallback(topic, qos);
   }
 }
 
 void TexturedMarkerPlugin::TopicEdited()
 {
   std::string topic = ui_.topic->text().trimmed().toStdString();
-  if (topic != topic_) {
+  connectCallback(topic, qos_);
+}
+
+void TexturedMarkerPlugin::connectCallback(const std::string& topic, const rmw_qos_profile_t& qos)
+{
+  ui_.topic->setText(QString::fromStdString(topic));
+
+  if ((topic != topic_) || !qosEqual(qos, qos_)) {
     initialized_ = false;
     markers_.clear();
     has_message_ = false;
@@ -150,22 +153,42 @@ void TexturedMarkerPlugin::TopicEdited()
     marker_arr_sub_.reset();
 
     topic_ = topic;
+    qos_ = qos;
     if (!topic.empty()) {
-      marker_arr_sub_ =
-        node_->create_subscription<marti_visualization_msgs::msg::TexturedMarkerArray>(
-        topic_,
-        rclcpp::QoS(1000),
-        std::bind(&TexturedMarkerPlugin::MarkerArrayCallback, this, std::placeholders::_1)
-        );
-      marker_sub_ = node_->create_subscription<marti_visualization_msgs::msg::TexturedMarker>(
-        topic_,
-        rclcpp::QoS(1000),
-        std::bind(&TexturedMarkerPlugin::MarkerCallback, this, std::placeholders::_1)
-      );
-
-      RCLCPP_INFO(node_->get_logger(), "Subscribing to %s", topic_.c_str());
+      auto known_topics = node_->get_topic_names_and_types();
+      if (known_topics.count(topic_) > 0) {
+        rclcpp::QoS topic_qos(rclcpp::QoSInitialization::from_rmw(qos_));
+        std::string topic_type = known_topics[topic_][0];
+        if (topic_type == "marti_visualization_msgs/msg/TexturedMarkerArray") {
+          marker_arr_sub_ =
+            node_->create_subscription<marti_visualization_msgs::msg::TexturedMarkerArray>(
+            topic_,
+            rclcpp::QoS(rclcpp::QoSInitialization::from_rmw(qos), qos),
+            std::bind(&TexturedMarkerPlugin::MarkerArrayCallback, this, std::placeholders::_1)
+            );
+          RCLCPP_INFO(node_->get_logger(), "Subscribing to %s", topic_.c_str());
+        }
+        else if(topic_type == "marti_visualization_msgs/msg/TexturedMarker") {
+          marker_sub_ = node_->create_subscription<marti_visualization_msgs::msg::TexturedMarker>(
+            topic_,
+            rclcpp::QoS(rclcpp::QoSInitialization::from_rmw(qos), qos),
+            std::bind(&TexturedMarkerPlugin::MarkerCallback, this, std::placeholders::_1)
+          );
+          RCLCPP_INFO(node_->get_logger(), "Subscribing to %s", topic_.c_str());
+        }
+        else {
+          RCLCPP_ERROR(node_->get_logger(),
+              "Unable to subscribe to topic %s (unsupported type %s).",
+              topic_.c_str(), topic_type.c_str());
+        }
+      }
+      else {
+        RCLCPP_ERROR(node_->get_logger(),
+            "Unable to subscribe to topic %s, (does not exist).", topic_.c_str());
+      }
     }
   }
+
 }
 
 void TexturedMarkerPlugin::ProcessMarker(const marti_visualization_msgs::msg::TexturedMarker marker)
@@ -506,9 +529,10 @@ void TexturedMarkerPlugin::Transform()
 
 void TexturedMarkerPlugin::LoadConfig(const YAML::Node & node, const std::string & path)
 {
+  LoadQosConfig(node, qos_);
   if (node["topic"]) {
-    std::string topic = node["topic"].as<std::string>();
-    ui_.topic->setText(boost::trim_copy(topic).c_str());
+    std::string topic = TrimString(node["topic"].as<std::string>());
+    ui_.topic->setText(topic.c_str());
   }
 
   TopicEdited();
@@ -516,7 +540,9 @@ void TexturedMarkerPlugin::LoadConfig(const YAML::Node & node, const std::string
 
 void TexturedMarkerPlugin::SaveConfig(YAML::Emitter & emitter, const std::string & path)
 {
-  emitter << YAML::Key << "topic" << YAML::Value <<
-    boost::trim_copy(ui_.topic->text().toStdString());
+  std::string topic = TrimString(ui_.topic->text().toStdString());
+  emitter << YAML::Key << "topic" << YAML::Value << topic;
+
+  SaveQosConfig(emitter, qos_);
 }
 }   // namespace mapviz_plugins
